big update. Level Up

+++ routes/lazada/rrs.js
@@ -2,7 +2,9 @@ var express = require('express');
 const router = express.Router();
 const rrsModel = require('../../models/RrsModel');
 const khoDuLieuModel = require("../../models/KhoDuLieuModel");
-
+const xlsx = require('node-xlsx');
+const fs = require("fs");
+const moment = require('moment');
 
 router.use((req, res, next) => {
     if (req.user) { req.owner = req.user.username; } else { req.owner = 'anonymous'; }
@@ -35,7 +37,7 @@ router.get('/', async function(req, res) {
 });
 
 router.get('/view', async (req,res) =>{
-    rrsData = await khoDuLieuModel.find();
+    rrsData = await khoDuLieuModel.find({owner: req.owner});
 
     
     
@@ -50,6 +52,34 @@ router.get('/view', async (req,res) =>{
     });
 
 });
+
+router.post('/export', async (req, res, next) => {
+    const d = new Date();
+    const today = d.getDate();
+    const currentMonth = d.getMonth() + 1; 
+
+  
+    const dataReg = await khoDuLieuModel.find({owner: req.owner}).exec();
+    var macs = [];
+    macs.push(["Username","Password", "Is Get", "Trạng Thái","Người Tạo", "Thời Gian Tạo"]);
+    for (var i = 0; i < dataReg.length; i++) {
+      macs.push([dataReg[i].username, dataReg[i].password,dataReg[i].isGet,dataReg[i].status,dataReg[i].owner,dataReg[i].created]);
+    }
+  
+    var datas = xlsx.build([
+      { name: "Excel", data: macs }
+    ]);
+  
+  
+    var time = moment().format('YYYYMMDDHHMMSS');
+      var file_path = '/download/' +"khoDuLieu_" +time + '.xlsx';
+      var write_path = './public' + file_path;
+      
+      console.log(write_path);
+      fs.writeFileSync(write_path, datas, {  });
+  
+      res.status(200).json({ ret_code: 0, success: true, file_path });
+  });
 
 router.post('/addData', async(req,res) =>{
     let newRrs = new rrsModel();
@@ -77,7 +107,7 @@ router.post('/addData', async(req,res) =>{
 });
 
 router.post('/deleteData', async(req,res)=>{
-    const result = await khoDuLieuModel.deleteMany();
+    const result = await khoDuLieuModel.deleteMany({owner: req.owner});
     console.log(result);
 
     res.status(200).json({
