+++ routes/remote/addScript.js
@@ -2,8 +2,13 @@ var express = require('express');
 const router = express.Router();
 const scriptModel = require('../../models/ScriptModel');
 
+router.use((req, res, next) => {
+  if (req.user) { req.owner = req.user.username; } else { req.owner = 'anonymous'; }
+  next();
+});
+
 router.get('/', async function (req, res) {
-  const listScript = await scriptModel.find().exec((err, result) => {
+  const listScript = await scriptModel.find({owner:req.owner}).exec((err, result) => {
     console.log(listScript);
     res.render('remote/addScript', {
       userData: req.user,
@@ -22,6 +27,7 @@ router.post('/addKichBan', async function (req, res) {
   newScript.scriptName = req.body.scriptName;
   newScript.duongDan = req.body.duongDan;
   newScript.scriptType = req.body.scriptType;
+  newScript.owner = req.owner;
   console.log(newScript);
 
   newScript.save().then((result) => {
@@ -34,9 +40,9 @@ router.post('/addKichBan', async function (req, res) {
 });
 
 router.post('/deleteKichBan', async (req, res) => {
-    const result = await scriptModel.deleteMany();
-  
-  
+    const result = await scriptModel.deleteMany({owner:req.owner,scriptType: req.body.scriptType});
+
+
     res.status(200).json({
       success: true,
       msg: 'Da xoa toan bo du lieu',
